Fix some timeout issues
This fixes some timeout issues. Peer permissions are valid for only 300 seconds and need to be refreshed before they expire. A new permission request for an existing IP is interpreted by TURN servers as a permission refresh request, so the `assert` was getting in the way of actually performing the refresh.

Secondly, the timeout event was set to occur at 10 / 9 of the expiry, causing dropped connections. I think this was intended to be 9 / 10. This would cause any refresh requests to occur just before the old request expires, leading to an uninterrupted connection.